video 11

+++ How_to_program_the_Best_Fit_Slope_p8.py
@@ -2,8 +2,11 @@ from statistics import mean
 import numpy as np
 import matplotlib.pyplot as plt
 
-xs=np.array([1,2,3,4,5,6],dtype=np.float64)
-ys=np.array([5,4,6,5,6,7],dtype=np.float64)
+# xs=np.array([1,2,3,4,5,6],dtype=np.float64)
+# ys=np.array([5,4,6,5,6,7],dtype=np.float64)
+
+xs=np.array([1,2,3,4,5,6])
+ys=np.array([5,4,6,5,6,7])
 
 def best_fit_slope(xs,ys):
     m=(((mean(xs)*mean(ys))-mean(xs*ys))
